Move batched_stream
