+++ dwds/debug_extension/web/background.dart
@@ -19,12 +19,8 @@ import 'package:collection/collection.dart' show IterableExtension;
 import 'package:dwds/data/devtools_request.dart';
 import 'package:dwds/data/extension_request.dart';
 import 'package:dwds/data/serializers.dart';
+import 'package:dwds/shared/batched_stream.dart';
 import 'package:dwds/src/sockets.dart';
-// NOTE(annagrin): using 'package:dwds/src/utilities/batched_stream.dart'
-// makes dart2js skip creating background.js, so we use a copy instead.
-// import 'package:dwds/src/utilities/batched_stream.dart';
-// Issue: https://github.com/dart-lang/sdk/issues/49973
-import 'package:dwds/src/web_utilities/batched_stream.dart';
 import 'package:js/js.dart';
 import 'package:js/js_util.dart' as js_util;
 import 'package:pub_semver/pub_semver.dart';
