+++ dwds/lib/shared/batched_stream.dart
