+++ dwds/web/client.dart
@@ -21,12 +21,8 @@ import 'package:dwds/data/extension_request.dart';
 import 'package:dwds/data/register_event.dart';
 import 'package:dwds/data/run_request.dart';
 import 'package:dwds/data/serializers.dart';
+import 'package:dwds/shared/batched_stream.dart';
 import 'package:dwds/src/sockets.dart';
-// NOTE(annagrin): using 'package:dwds/src/utilities/batched_stream.dart'
-// makes dart2js skip creating background.js, so we use a copy instead.
-// import 'package:dwds/src/utilities/batched_stream.dart';
-// Issue: https://github.com/dart-lang/sdk/issues/49973
-import 'package:dwds/src/web_utilities/batched_stream.dart';
 import 'package:js/js.dart';
 import 'package:sse/client/sse_client.dart';
 import 'package:uuid/uuid.dart';
